+++ docs/version-history.md
@@ -1,5 +1,10 @@
 # Version History
 
+- 1.1.0
+  - support laravel 10
+  - fixed `league()->byFixtureDate()`
+  - removed `prediction()->valueBetsByFixtureId()`
+  - removed `rival()->byTeamId()`
 - 1.0.1
   - bug fixed `setFilter`
 - 1.0.0